Refactor: replace monolithic entity code with ECS architecture
- Introduce EntityRegistry with index/generation EntityIDs and
  component storage via typed pools and bitset masks
- Add EventBus (observer pattern) for decoupled game events
- Add InputSystem with JSON-driven key bindings (arrow keys + WASD)
- Add AISystem, CollisionSystem, RenderSystem as pure system classes
- Add EntityFactory to create player, enemies, items, goal from layout
- Remove old procedural Player/Enemy/Item/Goal/Entity layer
- Maze now owns only tile data; entity spawning moved to EntityFactory
- All components implement to_json/from_json for future save/load

+++ MazeGame/src/rendering/VisualEffect.cpp
@@ -46,7 +46,7 @@ void VisualEffect::updateAll() {
         }
         else {
             fx.y -= fx.config.riseSpeed;
-            fx.alpha = 255 * (1.0f - static_cast<float>(elapsed) / fx.config.duration);
+            fx.alpha = 255.0f * (1.0f - static_cast<float>(elapsed) / static_cast<float>(fx.config.duration));
             ++i;
         }
     }
